refactor: 기존 인증 훅 제거 및 페이지 리팩토링
