+++ frontend/src/pages/SignInPage.tsx
@@ -5,9 +5,8 @@ import Button from '@/components/_common/Button/Button';
 import InputField from '@/components/_common/InputField/InputField';
 import Text from '@/components/_common/Text/Text';
 import { ROUTE_PATH } from '@/constants/routhPath';
-import useSignInMutation from '@/hooks/useSignInMutation';
+import { useSignInMutation } from '@/hooks/query/useAuthQuery';
 import useValidateInput from '@/hooks/useValidateInput';
-import logo from '@/assets/images/common/codelight-logo.svg';
 import { validateEmail, validatePassword } from '@/utils/authValidation';
 import { getKakaoAuthUrl } from '@/constants/oauth';
 
@@ -71,7 +70,6 @@ const SignInPage = () => {
   return (
     <>
       <div className='flex h-[calc(100dvh-56px)] flex-col items-center justify-center gap-4'>
-        <img src={logo} alt='Codelight Logo' className='mb-8 h-12' />
         <div className='relative flex w-[36.1rem] flex-col gap-0 rounded-2xl px-6'>
           <InputField onKeyDown={handleKeyDown}>
             <InputField.Label label='이메일' />
